v 1.0.0
This project marks my exciting journey into the world of Python programming. As my first venture, it embodies my initial steps into coding, showcasing my dedication to learning and applying Python's powerful capabilities. Through this project, I aim to not only enhance my programming skills but also to create a meaningful application that reflects my passion for technology and innovation. It's a significant milestone for me, representing the beginning of what I hope to be a rewarding and enriching coding adventure.